Add files via upload
This Python program does the following:

Fetch the data from the pages of the API.
Identify whether the response for each response-sources pair came from any of the sources.
List down the sources from which the response was formed. Returns an empty array if the response did not come from any source. The shortlisted sources will be called citations.
Return the citations for all objects coming from the API through a user-friendly UI.